requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==1.11.4
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-csv==2.1.0
idna==2.7
oauthlib==2.1.0
pkg-resources==0.0.0
PySocks==1.6.8
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
tweepy==3.6.0
unicodecsv==0.14.1
urllib3==1.23
